requirements.txt
pandas
faiss-cpu
sentence-transformers
fastapi
uvicorn
